# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/subscriptions/available": {
      "get": {
        "operationId": "SubscriptionsController_findAvailable",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Plans offered to the caller's organisation type.",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/me": {
      "get": {
        "operationId": "SubscriptionsController_myEntitlements",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "What the caller's organisation is currently allowed to do.",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions": {
      "post": {
        "operationId": "SubscriptionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionPlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Subscriptions"
        ]
      },
      "get": {
        "operationId": "SubscriptionsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/{id}": {
      "get": {
        "operationId": "SubscriptionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Subscriptions"
        ]
      },
      "patch": {
        "operationId": "SubscriptionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubscriptionPlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Subscriptions"
        ]
      },
      "delete": {
        "operationId": "SubscriptionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Subscriptions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Saveful for Business API",
    "description": "Food rescue platform API. All routes are prefixed `/api/v1`.\n\n**Authentication.** Every endpoint requires a bearer token unless marked otherwise. Obtain one from `POST /auth/login`, then send it as `Authorization: Bearer <token>`.\n\n**Enterprise endpoints** additionally require the organisation to be on the Enterprise plan, and are gated on the caller’s Enterprise role and scope. A 403 carrying `MISSING_PERMISSION` means the role is wrong; one carrying `OUTSIDE_SCOPE` means the role is right but the target sits outside what the caller can reach.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateSubscriptionPlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "maxSites": {
            "type": "number",
            "description": "Omit for unlimited (Enterprise).",
            "minimum": 1
          },
          "maxUserPerSite": {
            "type": "number",
            "description": "Omit for unlimited (Multi Site, Enterprise).",
            "minimum": 1
          },
          "priceMonthly": {
            "type": "number",
            "description": "Omit when the plan is quote-based — see contactSalesOnly.",
            "minimum": 0
          },
          "priceAnnual": {
            "type": "number",
            "minimum": 0
          },
          "priceMonthlyInr": {
            "type": "number",
            "description": "India (region = IN) pricing.",
            "minimum": 0
          },
          "priceAnnualInr": {
            "type": "number",
            "minimum": 0
          },
          "isPerSite": {
            "type": "boolean",
            "description": "Bills per location; the Stripe quantity tracks site count."
          },
          "contactSalesOnly": {
            "type": "boolean",
            "description": "Excluded from Checkout; the client routes to an enquiry form."
          },
          "applicableOrgTypes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "BUSINESS_SINGLE",
                "BUSINESS_MULTI",
                "CHARITY",
                "CHARITY_SINGLE",
                "CHARITY_MULTI",
                "FARMER_PRODUCER",
                "FARMER_CONSUMER"
              ]
            }
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isMostPopular": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "displayName",
          "features"
        ]
      },
      "UpdateSubscriptionPlanDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "maxSites": {
            "type": "number",
            "minimum": 1
          },
          "maxUserPerSite": {
            "type": "number",
            "minimum": 1
          },
          "priceMonthly": {
            "type": "number",
            "minimum": 0
          },
          "priceAnnual": {
            "type": "number",
            "minimum": 0
          },
          "priceMonthlyInr": {
            "type": "number",
            "minimum": 0
          },
          "priceAnnualInr": {
            "type": "number",
            "minimum": 0
          },
          "isPerSite": {
            "type": "boolean"
          },
          "contactSalesOnly": {
            "type": "boolean"
          },
          "applicableOrgTypes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "BUSINESS_SINGLE",
                "BUSINESS_MULTI",
                "CHARITY",
                "CHARITY_SINGLE",
                "CHARITY_MULTI",
                "FARMER_PRODUCER",
                "FARMER_CONSUMER"
              ]
            }
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isMostPopular": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}